requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
ariadne==0.20.1
async-timeout==4.0.2
attrs==23.2.0
backoff==2.2.1
biplist==1.0.3
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.98.0
filetype==1.2.0
flake8==6.0.0
frozenlist==1.4.1
fs==2.4.16
gql==3.4.1
graphql-core==3.2.3
h11==0.14.0
httptools==0.6.1
idna==3.7
iniconfig==2.0.0
iphone-backup-decrypt==0.7.0
mccabe==0.7.0
multidict==6.0.5
packaging==23.1
phonenumbers==8.13.15
pluggy==1.2.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.9
pyflakes==3.0.1
PyPika==0.48.9
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.22.0
uvloop==0.19.0; sys.platform != 'win32'
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.4
